% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastPATCH.R
\name{terraTILESpatches}
\alias{terraTILESpatches}
\title{Process raster tiles to create cleaned polygon patches}
\usage{
terraTILESpatches(input_dir, output_path, num_processes = NULL)
}
\arguments{
\item{input_dir}{Directory containing input raster tiles (.tif files)}

\item{output_path}{Output file path (supported formats: .shp, .gpkg, .geojson)}

\item{num_processes}{Number of parallel processes to use (NULL for automatic ~ max 10 cores)}
}
\value{
gpkg or geojson or shp plus an sf object containing the patched/clumped polygons (invisibly)
}
\description{
Process raster tiles to create cleaned polygon patches
}
